+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## [Released]
 
+## [1.6.0] - 2023-06-22
+
+- Sender identities by email
+
+## [1.5.2] - 2023-06-20
+
+- Removed uninitialized variable
+
 ## [1.5.1] - 2023-05-09
 
 - Added supports for base64 attachment